+++ src/components/story/StoryList.js
@@ -6,7 +6,7 @@ import {
   ProgressWrapper,
   Bar,
   ProgressBar,
-} from "components/sharedComponents/sharedComponents";
+} from "components/sharedComponents";
 import { PC_MIN, TABLET_MAX, TABLET_MIN } from "utils/media";
 
 const Container = styled.section`
