+++ src/containers/LogoContainer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Logo from "components/sharedComponents/Logo";
+import Logo from "components/Logo";
 const LogoContainer = ({ left,background }) => {
   return <Logo left={left} background={background}></Logo>;
 };
