+++ src/containers/storyPage/SlideContainer.js
