+++ src/pages/LinkPage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import LinkContainer from "containers/LinkContainer";
+import LinkContainer from "containers/linkPage/LinkContainer";
 
 const LinkPageWrapper = styled.section`
   width: 100%;
